GEN AI/04_Voice-to-Summary Meeting Assistant/requirements.txt
streamlit
openai-whisper
transformers
torch
sentencepiece
fpdf
wordcloud
matplotlib